requirements.txt
requests==2.27.0
SocksiPy_branch==1.01
urllib3==1.26.4